app/requirements.txt
pandas
pydantic
python-dotenv
langchain-openai
langchain-gigachat
sqlalchemy
streamlit
dotenv
jinja2
python-frontmatter
instructor
psycopg2
psycopg2-binary
langchain-ollama